[CurriculumPicker] Extract component
