+++ services/frontend/src/components/PopulationDetails/CourseTableModeSelector.jsx
@@ -1,0 +1,62 @@
+import { useEffect } from 'react'
+import { Form, Input, Radio } from 'semantic-ui-react'
+
+import { CurriculumPicker } from '@/components/material/CurriculumPicker'
+
+export const CourseTableModeSelector = ({
+  courseTableMode,
+  filteredStudents,
+  onStudentAmountLimitChange,
+  setCourseTableMode,
+  setCurriculum,
+  setStudentAmountLimit,
+  studentAmountLimit,
+  studyProgramme,
+  year,
+}) => {
+  useEffect(() => {
+    setStudentAmountLimit(Math.round(filteredStudents.length ? filteredStudents.length * 0.3 : 0))
+  }, [filteredStudents.length])
+
+  return (
+    <div style={{ display: 'flex', flexDirection: 'row' }}>
+      <div style={{ display: 'flex', flexDirection: 'column', gap: '26px' }}>
+        <Radio
+          checked={courseTableMode === 'curriculum'}
+          label="Choose curriculum"
+          name="coursesRadioGroup"
+          onChange={(_event, { value }) => setCourseTableMode(value)}
+          style={{ fontWeight: 'bold' }}
+          value="curriculum"
+        />
+        <Radio
+          checked={courseTableMode === 'all'}
+          label="Show all courses with at least"
+          name="coursesRadioGroup"
+          onChange={(_event, { value }) => setCourseTableMode(value)}
+          style={{ fontWeight: 'bold' }}
+          value="all"
+        />
+      </div>
+      <div>
+        <CurriculumPicker
+          disabled={courseTableMode !== 'curriculum'}
+          programmeCodes={[studyProgramme]}
+          setCurriculum={setCurriculum}
+          year={year}
+        />
+        <Form style={{ padding: '4px 4px 4px 8px' }}>
+          <Form.Field inline>
+            <Input
+              disabled={courseTableMode !== 'all'}
+              onChange={event => onStudentAmountLimitChange(event.target.value)}
+              style={{ width: '70px' }}
+              value={studentAmountLimit}
+            />
+            <label>total students</label>
+          </Form.Field>
+        </Form>
+      </div>
+    </div>
+  )
+}
